requirements.txt
# Core Dependencies
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
mysql-connector-python>=8.1.0

# GUI Dependencies
tkinter
tkcalendar>=1.6.1

# Image Processing
face-recognition>=1.3.0
dlib>=19.24.0

# Utilities
python-dateutil>=2.8.2
pandas>=2.0.0
matplotlib>=3.7.0

# Development Tools
pytest>=7.4.0
flake8>=6.1.0
black>=23.7.0

# Documentation
Sphinx>=7.1.0